requirements.txt
python-binance>=1.0.19
requests>=2.28.0
python-dotenv>=1.0.0
hypothesis>=6.0.0
pytest>=7.0.0
